move modals to component folder, car filter based on model
